+++ errors/handler_func.go
@@ -1,5 +1,13 @@
 package errors
 
+// HandlerFunc wraps a function and turns it into an error handler.
+type HandlerFunc func(err error)
+
+// Handle calls the underlying log function.
+func (h HandlerFunc) Handle(err error) {
+	h(err)
+}
+
 // HandlerLogFunc wraps a log-like function and turns it into an error handler.
 type HandlerLogFunc func(args ...interface{})