+++ errors/handler_func_test.go
@@ -9,6 +9,21 @@ import (
 
 func TestHandlerFunc(t *testing.T) {
 	var actual error
+	log := func(err error) {
+		actual = err
+	}
+
+	fn := errors.HandlerFunc(log)
+
+	expected := errors.New("error")
+
+	fn.Handle(expected)
+
+	assert.Equal(t, expected, actual)
+}
+
+func TestHandlerLogFunc(t *testing.T) {
+	var actual error
 	log := func(args ...interface{}) {
 		actual = args[0].(error)
 	}